+++ test/test_dmclock_server.cc
@@ -16,46 +16,9 @@
 #include "dmclock_util.h"
 #include "gtest/gtest.h"
 
-#include <dmtest-config.h>
-#ifdef HAVE_SYS_PRCTL_H
-#include <iostream>
-#include <sys/prctl.h>
-#include <errno.h>
+// process control to prevent core dumps during gtest death tests
+#include "dmcPrCtl.h"
 
-struct PrCtl {
-  int saved_state = -1;
-  int set_dumpable(int new_state) {
-    int r = prctl(PR_SET_DUMPABLE, new_state);
-    if (r) {
-      r = -errno;
-      std::cerr << "warning: unable to " << (new_state ? "set" : "unset")
-                << " dumpable flag: " << strerror(r)
-                << std::endl;
-    }
-    return r;
-  }
-  PrCtl(int new_state = 0) {
-    int r = prctl(PR_GET_DUMPABLE);
-    if (r == -1) {
-      r = errno;
-      std::cerr << "warning: unable to get dumpable flag: " << strerror(r)
-                << std::endl;
-    } else if (r != new_state) {
-      if (!set_dumpable(new_state)) {
-        saved_state = r;
-      }
-    }
-  }
-  ~PrCtl() {
-    if (saved_state < 0) {
-      return;
-    }
-    set_dumpable(saved_state);
-  }
-};
-#else
-struct PrCtl {};
-#endif
 
 namespace dmc = crimson::dmclock;
